feat: move util class to util package
- Moved InputValidationUtil class to util package
- Chaged InputValidationUtil package in PlayRacingCar class
